This begins support for the attr library
This was inspired by a bug report/feature request I saw on the jsl bug
tracker as well as a first stab on some features I have been
considering since the inceoption of this project.  This branch starts
out in a very unstable place as it might change as I consider better
ways to do this.